Tidying up basic layout
This is not a proper layout yet in that most elements are full width and we’d like want to show things side by side.

It at least adds styling to buttons and inputs, plus some minimal spacing which improves things quite a lot. A new Output component ensures code preview and copy is consistent between sections.