Adding basic table of contents
This is a functional implementation but it’s not ideal in that it requires the toc elements to defined manually/statically, leading to repetition.

It would be better if the contents were defined dynamically based on the components on the page. I did look into this briefly but it doesn’t seem there’s an elegant/simple way to do this. Ideally components would all emit a common event type with their name and anchor and the App component respond to these from all components to add to a common list held by the App comp.

This is something to come back to.