Adding basic dark mode support
This could definitely do with some more thought/refinement but as a first effort is much better than not having a dark mode.

In particular it feels like we now have lots of different colour combinations between borders, hovers, backgrounds etc. which I would like to standardise and rationalise to fewer.

If we get to that point it’s probably past when using a proper template for this app would make sense, which would hopefully have already thought about this.