Adding initial citation section
Uses the crosscite API to get a formatted citation based on the DOI to use as a default/initial value which can be manually edited.

This is currently using a fixed DOI and a live request, both of which needs to change as this new section evolves.

This section is based on the Freetext component which has been modified to be more modular/flexible.