Generating DOI citation locally
Adds a function accepting a set of values combined in the same format/style (Harvard APA) CrossCite returns.

Currently input values are static/fake set locally in the Citation component but should be taken from other components (some new components will need adding).

The mechanism for using the generated citation as the input for the citation freetext has been changed to become a trigger from a button - i.e. everytime the button is clicked the value will overrite whatever is in the freetext input.