Improving branding
Logo included in app header and flipped to look better visually.

I spent quite some time trying to get the header image to be constrained to the space needed by the header text (allowing it to resize dynamically). I couldn’t find a way to do this so for now it’s size is hard-coded. If we want to support mobile that will need reviewing, using breakpoint classes or similar.