Refactoring links into a Link component
As part of standardising classes and defining a better app structure. This is the first of a number of new base components that will form part of a wider app refactoring and tidyup.

For this component, I considered naming it ‘BaseLink’ or ‘LinkBase’ etc. but since I didn’t like either when used in other components (it suggested it needed scoping) I went with the simplier, more generic option. This is likely wrong and will need refactoring later but I don’t now why that will be yet so therefore don’t know what the more correct version would be.