Adding FormInput component
As part of creating new base components.

Inputs for radios and checkboxes don’t use this component as they currently aren’t styled, and applying these default styles may cause side effects. The aim of this refactoring is reduce the chance of visual inconsistencies due to partially updated or otherwise divergent classes. That isn’t a problem for these elements.