Fixing CI
The lint job was incorrectly set to only run for tags rather than opposite, now that commits from main are blocked it only needs to run on other branches.

We don’t need to build the app for releases as the tag points to an existing commit which will already have been deployed. This simplifies things as we don’t run the risk of deploying the same commit twice, rather we create a symlink from the commit to the tag and from the tag to the environment.

This means we don’t need to build or deploy the app for tags (only create symlinks).

The rules for the integration deploy need tightening to check a tag isn’t being deployed.

Symlinks need to use the `-n` mode to allow them to be overwritten correctly, otherwise they don’t get updated and we get stale references. See https://superuser.com/a/645847.