Adding first pass at tests
This mostly adds unit tests for components using vitest.

These are mostly successful, following the recommend principle that you don’t test the inner workings of the component but the rendered output and any events emitted or consumed.

However it has highlighted the need to refactor components to remove any ‘business logic’ such as date parsing and loading from JSON files, into pure functions in utils modules.

There are a small number of e2e tests needed to test clipboard interaction (as file inputs can’t be tested in vitest). These are essentially also unit tests needed for technical reasons.

I want to add proper e2e tests later, filling in a set of test data and checking the citation preview is rendered correctly. Whilst this only tests a subset of elements, the others are used standalone so are fine to be covered by unit tests.