Updating identifiers based on title (scheme)
Previously we updated based on the identifier value but this means if the value changes (as Esri values can), we’d never match the previous identifier (as its value will be differenent).

Since we (currently) only allow a single instance of each identifier scheme, we use that as a stable, but still unique, reference instead.