Adding version and stability to section components
This is both an internal change (files renamed) and user-facing with properties defined in section titles.

All components will start as version 1.0 and a stability of ‘stable’. Future changes will increase the version and if needed change to ‘experimental’.

Normally we would only have a single version of each component but as this app will sometimes be used to test different implementations of the same thing, having a structured way to do that seems sensible. This wouldn’t be the case in most other apps I imagine.