Refactoring epilogue component
Uses a utils method to get the build environment info and pass this as a prop (for easier testing of possible states).

This commits begins removing version information from component files names, as this was too annoying to work with and we didn’t keep the old components anyway. Versioning will be kept as a property of each (section) component.