Refactoring Table of Contents component
We no longer need to maintain a separate list of TOC items, which could get out of sync with components included in the page.

The inbuilt Teleport component is used to render the TOC link for each section header into the TOC component, which is now effectively an empty container.

This removes the need for handling events (which didn’t work anyway) and feels a cleaner design overall.