Refactoring how publisher/distributor are determined
Previously we used the DOI identifier as a proxy for PDC with that resting on whether it was a dataset.

Now that we have closed licences and access that no longer holds (if either of those are true PDC won’t apply).

Based on that and the resource type, there is now a very crude flowchart for determining who the distributor is (which almost always is also the publisher) - see https://gitlab.data.bas.ac.uk/MAGIC/data-management/-/issues/41.

This logic has been implemented into fucntions to get the distributor and publisher, from which decisions can be made on what to show.

For example the DOI is now only shown if the publisher is the PDC, and the citation publisher is no longer determined only by whether there’s a DOI or not.

This general logic is too simplistic and will need refining in the future but is better than what we had.

This has required a few general changes:
- disabled buttons
- licence emit event
- the Esri identifier is currently always shown (as we don’t have proper logic to say when it shouldn’t)