Squashed commit of the following:
Additions:

- basic 2D and 3D ArcGIS JS SDK components for showing the selected Well Known Extent as a a map and globe
- basic Esri OAuth implementation to access restricted feature service holding densified polygons for each WKE
- support functions for selecting the basemap and UI theme based on dark mode
- test mocks to workaround inability to test ArcGIS JS SDK with vitest
- Python support code to create densified polygons for each WKE and create/populate a feature service in AGOL

This work is a proof of concept and minimally viable (or just below in some areas) and will need significant improvements, fixed and refactoring. In particular issues with layer visibility, both in setting from Python and when visualising, causing features not to show when expected.

As the ArcGIS SDK uses lots of client side code, dynamically imported at runtime, standard tree-shaking techniques are
not possible and the bundle size for this app is now considerably larger (requiring GL runners resources to be increased). However as the code is chunked, end-users loading the app should not need to load the vast majority of it.

The lack of testing support for this code, which is inherently more complex, is a significant weakness. Whilst we can't (for the time being) use unit tests, I am thinking about what would make sense in terms of e2e tests (the logical test is to inspect the map and check that a single feature is shown and it's in view but that's not straight forward to do).