Adding external services section
Initially only for Esri but in time we’ll have other external services such as Office 365.

Moves and refactors the initial code from the extent globe.

As the functions for generating and parsing the Esri auth token don’t reply on the ArcGIS library we can test them. Therefore I’ve put them in a seaprate utils module.

To make the output of the Esri token easier to understand I’ve added a computed datetime rather than an offset in seconds.