Adding basic staging for file downloads
Adds a way to upload file based downloads to a staging location such that they have a URL.

This is the first step to adding other download types such as SAN or OneDrive based files (that already have URLs).

Initially a very basic backend upload script is used as designing how the app iteracts with that is more important.

Currently e2e tests use this live endpoint directly. This may change if it causes reliability problems.