Allow ‘all rights reserved’ licence for unrestricted items
Currently we consider the All Rights Reserved licence a closed licence - which means it can’t be selected for a public map.

Since we added this licence for open (published) maps that’s a problem. In the long term we should probably remove the open property and replace with an access type enum or similar.

For now it’s easier to set the open property to true.